Added some validation + boilerplate for sequencing
Start and cancellation of games are now validated, checking that the
initiating user is actually the owner of the game. Also refactored this
a bit, removing code duplication.

Also added some boiler plate for the upcoming game mechanics, and
fleshed out PlayerData with respect to game start and player sequencing,
since the 'seating order' should be determined randomly, and stay the
same throughout. Also added some provision for players randomly leaving
the game mid-round. The thought is that in such a case, the round is
re-started, dealing out roles anew.